+++ src/blocks/falling.h
@@ -42,9 +42,9 @@ class BlockFalling: public BlockBasic
 public:
   bool affectedBlock(int block);
 
-  void onNeighbourBroken(User* user, int8_t oldblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction);
-  bool onPlace(User* user, int8_t newblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction);
-  void onNeighbourMove(User* user, int8_t newblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction);
+  void onNeighbourBroken(User* user, int16_t oldblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction);
+  bool onPlace(User* user, int16_t newblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction);
+  void onNeighbourMove(User* user, int16_t newblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction);
 private:
   void applyPhysics(User* user, int32_t x, int8_t y, int32_t z, int map);
 };
