+++ src/blocks/ladder.h
@@ -84,7 +84,7 @@ public:
       @param z The z position of the current neighbour block being called
       @param direction The direction of the neighbour block that was broken
     */
-   void onNeighbourBroken(User* user, int8_t oldblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction);
+   void onNeighbourBroken(User* user, int16_t oldblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction);
 
    /** Fired when a block is placed
       @param user The user who placed the block
@@ -94,7 +94,7 @@ public:
       @param z The z position of where the block was placed
       @param direction The direction that the user is facing
     */
-   bool onPlace(User* user, int8_t newblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction);
+   bool onPlace(User* user, int16_t newblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction);
 
    /** Fired when a neighbour block is placed
       @param user The user who placed the block
@@ -108,7 +108,7 @@ public:
       allows another block's onReplace method to either allow of deny the replacement
       of a block with the block you're replacing it with.
     */
-   void onNeighbourPlace(User* user, int8_t newblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction);
+   void onNeighbourPlace(User* user, int16_t newblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction);
 
    /** Fired when a block is to be placed
       @param user The user who placed the block
@@ -122,7 +122,7 @@ public:
       the block change the block to be BLOCK_AIR and check for BLOCK_AIR
       in the onPlace callback.
     */
-   void onReplace(User* user, int8_t newblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction);
+   void onReplace(User* user, int16_t newblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction);
 
    /** Fired when a neighbour block is moving position.
       @param user The user who broke the block
@@ -132,6 +132,6 @@ public:
       @param z The z position of the current neighbour block being called
       @param direction The direction that the block was in
     */
-   void onNeighbourMove(User* user, int8_t oldblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction);
+   void onNeighbourMove(User* user, int16_t oldblock, int32_t x, int8_t y, int32_t z, int map, int8_t direction);
 };
